+++ lib/application/stats/stats.dart
@@ -4,13 +4,24 @@
 // Dart imports:
 import 'dart:async';
 
+// Flutter imports:
+import 'package:flutter/material.dart';
+
+// Package imports:
+import 'package:get_it/get_it.dart';
+import 'package:screen_retriever/screen_retriever.dart';
+
 // Project imports:
+import 'package:da_kanji_mobile/application/screensaver/screensaver.dart';
+import 'package:da_kanji_mobile/entities/settings/settings.dart';
+import 'package:da_kanji_mobile/entities/user_data/user_data.dart';
+import 'package:da_kanji_mobile/globals.dart';
 import 'package:da_kanji_mobile/repositories/analytics/event_logging.dart';
 
 // Package imports:
 
-// Project imports:
-import 'package:da_kanji_mobile/entities/user_data/user_data.dart'; 
+
+
 
 /// Tracking of statistics and communication with the local SQL Stas DB
 /// Also tracks daily and monthly usage
@@ -34,6 +45,13 @@ class Stats{
   late Timer saveStatsTimer;
   /// After how many seconds should the stats be saved to disk
   final int saveStatsTimerInterval = 60;
+  /// The position of the cursor last tick (updates every
+  /// `updateStatsTimerInterval` seconds)
+  Offset lastCursorPosition = Offset.zero;
+  /// Timer that starts when the cursor didnt move for one tick (see
+  /// `lastCursorPosition`). When the timer finishes, starts a screensaver
+  Timer? screenSaverTimer;
+
 
   /// Before using this `init()` needs to be called
   Stats(
@@ -47,6 +65,24 @@ class Stats{
       Duration(seconds: updateStatsTimerInterval), (timer) async {
         if(appActive){
           await updateStats();
+        }
+
+        // on desktop keep track if the cursor moves and if not
+        // show a screensaver after the user defined amount of time
+        if(g_desktopPlatform && GetIt.I<Settings>().wordLists.autoStartScreensaver){
+          Offset pos = await screenRetriever.getCursorScreenPoint();
+          if(lastCursorPosition != pos){
+            lastCursorPosition = pos;
+            screenSaverTimer?.cancel();
+            screenSaverTimer = null;
+          }
+          else {
+            screenSaverTimer ??= Timer(
+              Duration(seconds: GetIt.I<Settings>().wordLists.screenSaverSecondsToStart),
+              () async =>
+                startScreensaver(GetIt.I<Settings>().wordLists.screenSaverWordLists)
+            );
+          }
         }
       }
     );