+++ lib/application/word_lists/anki.dart
@@ -1,5 +1,10 @@
+// Dart imports:
 import 'dart:io';
 
+// Package imports:
+import 'package:get_it/get_it.dart';
+
+// Project imports:
 import 'package:da_kanji_mobile/application/anki/anki.dart';
 import 'package:da_kanji_mobile/entities/anki/anki_note.dart';
 import 'package:da_kanji_mobile/entities/settings/settings.dart';
@@ -7,9 +12,6 @@ import 'package:da_kanji_mobile/entities/tree/tree_node.dart';
 import 'package:da_kanji_mobile/entities/word_lists/word_lists_data.dart';
 import 'package:da_kanji_mobile/entities/word_lists/word_lists_queries.dart';
 import 'package:da_kanji_mobile/entities/word_lists/word_lists_sql.dart';
-import 'package:get_it/get_it.dart';
-
-
 
 /// Renders each word list entry to an image and stores it in the temp directory
 /// returns a List with all the [File]s created