+++ lib/entities/settings/settings_word_lists.dart
@@ -1,10 +1,12 @@
 // Flutter imports:
-import 'package:collection/collection.dart';
-import 'package:da_kanji_mobile/entities/iso/iso_table.dart';
 import 'package:flutter/material.dart';
 
 // Package imports:
+import 'package:collection/collection.dart';
 import 'package:json_annotation/json_annotation.dart';
+
+// Project imports:
+import 'package:da_kanji_mobile/entities/iso/iso_table.dart';
 
 part 'settings_word_lists.g.dart';
 
@@ -105,6 +107,67 @@ class SettingsWordLists with ChangeNotifier {
   /// When exporting to PDF, should kana be included
   set pdfIncludeKana (bool value) {
     _pdfIncludeKana = value;
+    notifyListeners();
+  }
+
+
+  /// The default value for `screenSaverSecondsToStart`
+  @JsonKey(includeFromJson: false, includeToJson: false)
+  // ignore: constant_identifier_names
+  static const bool d_autoStartScreensaver = false;
+  /// Should the screen saver start automatically after the set time
+  @JsonKey(defaultValue: d_autoStartScreensaver)
+  bool _autoStartScreensaver = d_autoStartScreensaver;
+  /// Should the screen saver start automatically after the set time
+  bool get autoStartScreensaver => _autoStartScreensaver;
+  /// Should the screen saver start automatically after the set time
+  set autoStartScreensaver (bool value) {
+    _autoStartScreensaver = value;
+    notifyListeners();
+  }
+
+  /// The default value for `screenSaverWordList`
+  @JsonKey(includeFromJson: false, includeToJson: false)
+  // ignore: constant_identifier_names
+  static const List<int> d_screenSaverWordLists = [];
+  /// The word lists that should be included in the screen saver
+  @JsonKey(defaultValue: d_screenSaverWordLists)
+  List<int> _screenSaverWordLists = d_screenSaverWordLists;
+  /// The word lists that should be included in the screen saver
+  List<int> get screenSaverWordLists => _screenSaverWordLists;
+  /// The word lists that should be included in the screen saver
+  set screenSaverWordLists (List<int> value) {
+    _screenSaverWordLists = value;
+    notifyListeners();
+  }
+
+  /// The default value for `screenSaverSecondsToStart`
+  @JsonKey(includeFromJson: false, includeToJson: false)
+  // ignore: constant_identifier_names
+  static const int d_screenSaverSecondsToStart = 60*2;
+  /// How many seconds until the screen saver starts
+  @JsonKey(defaultValue: d_screenSaverSecondsToStart)
+  int _screenSaverSecondsToStart = d_screenSaverSecondsToStart;
+  /// How many seconds until the screen saver starts
+  int get screenSaverSecondsToStart => _screenSaverSecondsToStart;
+  /// How many seconds until the screen saver starts
+  set screenSaverSecondsToStart (int value) {
+    _screenSaverSecondsToStart = value;
+    notifyListeners();
+  }
+
+  /// The default value for `screenSaverSecondsToNextCard`
+  @JsonKey(includeFromJson: false, includeToJson: false)
+  // ignore: constant_identifier_names
+  static const int d_screenSaverSecondsToNextCard = 10;
+  /// How many seconds one card should be visible in screen saver mode
+  @JsonKey(defaultValue: d_screenSaverSecondsToNextCard)
+  int _screenSaverSecondsToNextCard = d_screenSaverSecondsToNextCard;
+  /// How many seconds one card should be visible in screen saver mode
+  int get screenSaverSecondsToNextCard => _screenSaverSecondsToNextCard;
+  /// How many seconds one card should be visible in screen saver mode
+  set screenSaverSecondsToNextCard (int value) {
+    _screenSaverSecondsToNextCard = value;
     notifyListeners();
   }