+++ lib/repositories/anki/anki_android.dart
@@ -1,11 +1,10 @@
-import 'package:da_kanji_mobile/repositories/anki/anki_data.dart';
+// Package imports:
 import 'package:flutter_ankidroid/flutter_ankidroid.dart';
 
 // Project imports:
 import 'package:da_kanji_mobile/entities/anki/anki_note.dart';
 import 'package:da_kanji_mobile/entities/settings/settings_anki.dart';
-
-
+import 'package:da_kanji_mobile/repositories/anki/anki_data.dart';
 
 /// Class to communicate with AnkiDroid (anki on android)
 class AnkiAndroid {