+++ lib/repositories/anki/anki_desktop.dart
@@ -12,8 +12,6 @@ import 'package:da_kanji_mobile/entities/anki/anki_note.dart';
 import 'package:da_kanji_mobile/entities/settings/settings_anki.dart';
 import 'package:da_kanji_mobile/repositories/anki/anki_data.dart';
 
-
-
 /// Class to communicate with anki desktop
 class AnkiDesktop {