+++ lib/widgets/word_lists/word_list_node.dart
@@ -1,28 +1,29 @@
 // Flutter imports:
-import 'package:da_kanji_mobile/application/word_lists/anki.dart';
-import 'package:da_kanji_mobile/application/word_lists/csv.dart';
-import 'package:da_kanji_mobile/application/word_lists/images.dart';
-import 'package:da_kanji_mobile/entities/user_data/user_data.dart';
-import 'package:da_kanji_mobile/entities/word_lists/word_lists_sql.dart';
-import 'package:da_kanji_mobile/widgets/anki/anki_not_setup_dialog.dart';
-import 'package:da_kanji_mobile/widgets/widgets/loading_popup.dart';
-import 'package:file_picker/file_picker.dart';
 import 'package:flutter/material.dart';
 
 // Package imports:
 import 'package:easy_localization/easy_localization.dart';
+import 'package:file_picker/file_picker.dart';
 import 'package:get_it/get_it.dart';
-import 'package:pdf/widgets.dart' as pw;
 import 'package:path/path.dart' as p;
+import 'package:pdf/widgets.dart' as pw;
+import 'package:universal_io/io.dart';
 
 // Project imports:
+import 'package:da_kanji_mobile/application/screensaver/screensaver.dart';
+import 'package:da_kanji_mobile/application/word_lists/anki.dart';
+import 'package:da_kanji_mobile/application/word_lists/csv.dart';
+import 'package:da_kanji_mobile/application/word_lists/images.dart';
 import 'package:da_kanji_mobile/application/word_lists/pdf.dart';
 import 'package:da_kanji_mobile/entities/tree/tree_node.dart';
+import 'package:da_kanji_mobile/entities/user_data/user_data.dart';
 import 'package:da_kanji_mobile/entities/word_lists/word_list_types.dart';
 import 'package:da_kanji_mobile/entities/word_lists/word_lists_data.dart';
+import 'package:da_kanji_mobile/entities/word_lists/word_lists_sql.dart';
 import 'package:da_kanji_mobile/globals.dart';
 import 'package:da_kanji_mobile/locales_keys.dart';
-import 'package:universal_io/io.dart';
+import 'package:da_kanji_mobile/widgets/anki/anki_not_setup_dialog.dart';
+import 'package:da_kanji_mobile/widgets/widgets/loading_popup.dart';
 
 /// All actions a user can do when clicking the 
 enum  WordListNodePopupMenuButtonItems {
@@ -32,6 +33,7 @@ enum  WordListNodePopupMenuButtonItems {
   toImages,
   toPdf,
   toCSV,
+  useAsScreenSaver,
 }
 
 /// One Node of the word lists tree, can either be a folder or a word list
@@ -320,6 +322,9 @@ class _WordListNodeState extends State<WordListNode> {
                             case WordListNodePopupMenuButtonItems.toCSV:
                               toCSVPressed();
                               break;
+                            case WordListNodePopupMenuButtonItems.useAsScreenSaver:
+                              startScreensaver([widget.node.id]);
+                              break;
                           }
                         },
                         itemBuilder: (context) => [
@@ -363,6 +368,12 @@ class _WordListNodeState extends State<WordListNode> {
                                   LocaleKeys.WordListsScreen_export_csv.tr()
                                 )
                               ),
+                              PopupMenuItem(
+                                value: WordListNodePopupMenuButtonItems.useAsScreenSaver,
+                                child: Text(
+                                  LocaleKeys.WordListsScreen_screensaver_use_as.tr()
+                                ),
+                              )
                             ]
                         ],
                       ),
@@ -528,4 +539,5 @@ class _WordListNodeState extends State<WordListNode> {
 
   }
 
+  
 }